updated query params

+++ api/playlistRouter.js
@@ -15,22 +15,15 @@ import {
 } from "#db/query/playlist_songs";
 import { getSongById } from "#db/query/songs";
 import { requireUser } from "#middleware/requireUser";
+import requireQuery from "#middleware/requireQuery";
 
 const router = express.Router();
 
-router.use(requireUser);
-
 router.get("/", async (req, res) => {
-  res.status(200).json(await getPlaylists(req.user.id));
+  res.status(200).json(await getPlaylists(req.query));
 });
 
-router.get("/:query/:limit/:offset", async (req, res) => {
-  const { query, limit, offset } = req.query;
-
-  res.status(200).json(await getPlaylistsByName(query, limit, offset));
-});
-
-router.post("/", requireBody(["name"]), async (req, res) => {
+router.post("/", requireUser, requireBody(["name"]), async (req, res) => {
   const playlist = await createPlaylist({ ...req.body, owner_id: req.user.id });
 
   if (!playlist) {
@@ -46,7 +39,7 @@ router.param("id", async (req, res, next, id) => {
   if (!playlist)
     return res.status(404).send("Couldn't find a playlist with that id.");
 
-  if (playlist.owner_id !== req.user.id)
+  if (!playlist.is_public && playlist.owner_id !== req.user?.id)
     return res.status(401).send("Access denied.");
 
   req.playlist = playlist;
@@ -58,10 +51,16 @@ router
   .get(async (req, res) => {
     res.status(200).json(req.playlist);
   })
-  .delete(async (req, res) => {
+  .delete(requireUser, async (req, res) => {
+    if (req.playlist.owner_id !== req.user?.id)
+      return res.status(401).send("Access denied.");
+
     res.status(204).json(await deletePlaylist(req.playlist.id));
   })
-  .put(async (req, res) => {
+  .put(requireUser, async (req, res) => {
+    if (req.playlist.owner_id !== req.user?.id)
+      return res.status(401).send("Access denied.");
+
     const playlist = await updatePlaylist(req.playlist.id, req.body);
     res.status(200).json(playlist);
   });
@@ -84,26 +83,33 @@ router.param("songId", async (req, res, next, id) => {
   next();
 });
 
-router.post("/:id/songs/:songId", async (req, res) => {
-  const playlist_song = await addToPlaylist({
-    playlist_id: req.playlist.id,
-    song_id: req.song.id,
+router
+  .route("/:id/songs/:songId")
+  .post(async (req, res) => {
+    if (req.playlist.owner_id !== req.user?.id)
+      return res.status(401).send("Access denied.");
+
+    const playlist_song = await addToPlaylist({
+      playlist_id: req.playlist.id,
+      song_id: req.song.id,
+    });
+
+    res.status(200).json(playlist_song);
+  })
+  .delete(async (req, res) => {
+    if (req.playlist.owner_id !== req.user?.id)
+      return res.status(401).send("Access denied.");
+
+    const playlist_song = await removeFromPlaylist({
+      playlist_id: req.playlist.id,
+      song_id: req.song.id,
+    });
+
+    if (!playlist_song) {
+      return res.status(404).send("Couldn't find a song to delete.");
+    }
+
+    res.status(204);
   });
-
-  res.status(200).json(playlist_song);
-});
-
-router.delete("/:id/songs/:songId", async (req, res) => {
-  const playlist_song = await removeFromPlaylist({
-    playlist_id: req.playlist.id,
-    song_id: req.song.id,
-  });
-
-  if (!playlist_song) {
-    return res.status(404).send("Couldn't find a song to delete.");
-  }
-
-  res.status(204);
-});
 
 export default router;
